Refactored the file storage service file into a module.
