+++ src/services/fileStorage/mockFileStorage.service.ts
@@ -1,0 +1,15 @@
+import { FileStorage } from ".";
+
+export class MockFileStorage implements FileStorage {
+  storeFile = async (fileBuffer: Buffer, fileId: string): Promise<string> => {
+    return `https://mockstorage.com/${fileId}`;
+  };
+
+  getFileId = async (filePath: string): Promise<string> => {
+    return "mockId";
+  };
+
+  deleteFile = async (fileName: string): Promise<void> => {
+    return;
+  };
+};
